requirements.txt
fastmcp==2.14.5
httpx==0.28.1
requests==2.32.5
uvicorn==0.40.0
azure-data-tables==12.5.0